Fix Discord ID db storage
This commit fixes a bug wherein Discord IDs would fail to store in the
production Postgres database, even though they stored properly in the
development SQLite database. SQLite by default allows integers to
surpass Django's internal limit for IntegerField values, while
Postgres does not, so it was failing to store the Discord IDs, which
surpass the limit. By updating the discord_id field to use
PositiveBigIntegerField, Postgres allocates more space for the
Discord IDs and allows them to be stored.

+++ configs/models.py
@@ -11,7 +11,7 @@ class Config(models.Model):
         on_delete=models.CASCADE,
         primary_key=True,
     )
-    discord_id = models.IntegerField(
+    discord_id = models.PositiveBigIntegerField(
         blank=True,
         default=0,
         unique=True,
